Fix a lot of warnings when compiling with -Wall + GCC 7.2.0

+++ Misc/TextHelper.cpp
@@ -70,7 +70,7 @@ namespace OSHGui
 		//---------------------------------------------------------------------------
 		void TextHelper::Remove(int index, int length)
 		{
-			if (index >= text_.length())
+			if (index >= static_cast<int>(text_.length()))
 			{
 				return;
 			}
